Split out credentials.js
