+++ src/components/DemoConfigurer.vue
@@ -83,13 +83,13 @@ export default {
       'roleName',
     ]),
     ...mapState({
-      role: state => state.auth.role,
-      availableRoles: state => state.auth.availableRoles,
-      activeRoleTab: state => state.auth.activeRoleTab,
-      accessToken: state => state.auth.accessToken,
-      clientId: state => state.auth.clientId,
-      clientSecret: state => state.auth.clientSecret,
-      userToken: state => state.auth.userToken,
+      role: state => state.role.role,
+      availableRoles: state => state.role.availableRoles,
+      activeRoleTab: state => state.role.activeRoleTab,
+      accessToken: state => state.credentials.accessToken,
+      clientId: state => state.credentials.clientId,
+      clientSecret: state => state.credentials.clientSecret,
+      userToken: state => state.credentials.userToken,
     }),
   },
   methods: {
